refactoring

+++ src/main/java/com/example/onlinebank/transaction_service/service/TransactionService.java
@@ -4,10 +4,14 @@ import com.example.onlinebank.account_service.entity.Account;
 import com.example.onlinebank.account_service.exception.AccountNotFoundException;
 import com.example.onlinebank.account_service.repository.AccountRepository;
 import com.example.onlinebank.payment_service.entity.Payment;
+import com.example.onlinebank.payment_service.entity.PaymentStatus;
+import com.example.onlinebank.payment_service.exception.PaymentCreationException;
+import com.example.onlinebank.payment_service.exception.PaymentNotException;
 import com.example.onlinebank.payment_service.service.PaymentService;
 import com.example.onlinebank.transaction_service.entity.Transaction;
 import com.example.onlinebank.transaction_service.exception.TransactionNotFoundException;
 import com.example.onlinebank.transaction_service.repository.TransactionRepository;
+import lombok.SneakyThrows;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
@@ -51,13 +55,18 @@ public class TransactionService {
         return transactionRepository.save(transaction);
     }
 
+    @SneakyThrows
     public Transaction createTransactionWithPayment(Long accountId, BigDecimal amount, String description) {
         Transaction transaction = createTransaction(accountId, amount, description);
         Payment payment = new Payment();
         payment.setTransaction(transaction);
         payment.setAmount(amount);
-        payment.setPaymentStatus("NEW");
-        paymentService.createPayment(payment);
+        payment.setPaymentStatus(PaymentStatus.NEW);
+        try {
+            paymentService.createPayment(payment);
+        }catch (Exception e) {
+            throw new PaymentCreationException("Payment creation failed");
+        }
         return transaction;
     }
 
@@ -69,19 +78,31 @@ public class TransactionService {
         return transactionRepository.save(transaction);
     }
 
-    public void updateTransactionAndPaymentStatus(Long transactionId, String paymentStatus) {
+    @SneakyThrows
+    public void updateTransactionAndPaymentStatus(Long transactionId, PaymentStatus paymentStatus) {
         Transaction transaction = getTransaction(transactionId);
+        if (transaction.getPayment() == null) {
+            throw new PaymentNotException("Payment not found for transaction");
+        }
         Payment payment = paymentService.getPayment(transaction.getPayment().getId());
         paymentService.updatePaymentStatus(payment.getId(), paymentStatus);
     }
 
+    @SneakyThrows
     public void deleteTransactionAndPayment(Long transactionId) {
         Transaction transaction = getTransaction(transactionId);
+        if (transaction.getPayment() == null) {
+            throw new PaymentNotException("Payment not found for transaction");
+        }
         paymentService.deletePayment(transaction.getPayment().getId());
         deleteTransaction(transactionId);
     }
 
     public void deleteTransaction(Long transactionId) {
+        Transaction transaction = getTransaction(transactionId);
+        if (transaction == null) {
+            throw new TransactionNotFoundException("Transaction not found");
+        }
         transactionRepository.deleteById(transactionId);
     }
 }
